api recommended

+++ app/src/main/java/studio/stilip/proffer/app/fragments/ads/category/spares/car/CategorySparesCarFragment.kt
@@ -62,7 +62,7 @@ class CategorySparesCarFragment : Fragment(R.layout.fragment_category_spares_car
                 }
 
                 val name = editName.text.toString()
-                val price = editPrice.text.toString().toInt()
+                val price = editPrice.text.toString().toDouble()
                 val phone = editPhone.text.toString()
                 val description = editDescription.text.toString()
                 val address = editLocation.text.toString()
